% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{beta_coordinate_descent_R_maxLik}
\alias{beta_coordinate_descent_R_maxLik}
\title{Wrapper function for beta optimization using coordinate ascend based on Newton–Raphson update}
\usage{
beta_coordinate_descent_R_maxLik(
  beta_cd,
  beta0_cd,
  nu_cd,
  ga_cd,
  betaPRE,
  t0,
  t1,
  Y_cd,
  X_cd,
  theta_cd,
  maX_cd_iter,
  tol
)
}
\arguments{
\item{beta_cd}{Initial beta parameter}

\item{beta0_cd}{Beta0 parameter}

\item{nu_cd}{Nu parameter}

\item{ga_cd}{Gamma parameter}

\item{betaPRE}{Previous beta value}

\item{t0}{the "spike" hyperparameter in beta's spike-and-slab prior}

\item{t1}{the "slab" hyperparameter in beta's spike-and-slab prior}

\item{Y_cd}{Response vector}

\item{X_cd}{Predictor matrix}

\item{theta_cd}{Theta parameter (default: number of predictors)}
}
\value{
Optimized beta parameter
}
\description{
Wrapper function for beta optimization using coordinate ascend based on Newton–Raphson update
}
