% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{CiS_j_fun}
\alias{CiS_j_fun}
\title{Calculate CiS for a single index}
\usage{
CiS_j_fun(
  test_index,
  beta_opt,
  beta0_opt,
  nu_opt,
  ga_opt,
  dataXY,
  add_correc_CiS = 0.001
)
}
\arguments{
\item{test_index}{Index to test}

\item{beta_opt}{Optimized beta vector}

\item{beta0_opt}{Optimized beta0 value}

\item{nu_opt}{Optimized nu value}

\item{ga_opt}{Optimized gamma value}

\item{dataXY}{List containing data X and Y}

\item{add_correc_CiS}{Correction factor for CiS (default: 0.001)}
}
\value{
CiS value
}
\description{
Calculate CiS for a single index
}
