% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{curvature_error}
\alias{curvature_error}
\title{Calculate curvature error}
\usage{
curvature_error(err, nu_cur, ga_cur)
}
\arguments{
\item{err}{Error vector}

\item{nu_cur}{Nu parameter for curvature}

\item{ga_cur}{Gamma parameter for curvature}
}
\value{
Vector of curvature errors
}
\description{
Calculate curvature error
}
