% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{TVS}
\alias{TVS}
\title{Calculate TVS values}
\usage{
TVS(
  dataXY,
  init_beta_TVS = rep(1, 8),
  B = 300,
  sig_cutoff = 0.05,
  init_beta0_TVS = 1,
  init_nu_TVS = 1,
  init_gamma_TVS = 1,
  init_theta_TVS = 0.5,
  SS_t0_TVS = 10,
  SS_t1_TVS = 1,
  hyper_mu_beta0_TVS = 0,
  hyper_sigma_beta0_TVS = 1e+06,
  hyper_mu_nu_TVS = 1,
  hyper_sigma_nu_TVS = 1,
  hyper_c_gamma_TVS = 1e-04,
  hyper_d_gamma_TVS = 1e-04,
  hyper_a_theta_TVS = 1,
  hyper_b_theta_TVS = -1,
  max_iter_TVS = 100,
  tol_TVS = 1e-06,
  add_correc_CiS = 0.001
)
}
\arguments{
\item{dataXY}{List containing data X and Y}

\item{init_beta_TVS}{Initial vector for beta}

\item{B}{Number of iterations (default: 300)}

\item{sig_cutoff}{Threshold used for significance cutoff in variable selection (default: 0.05)}

\item{init_beta0_TVS}{Initial value for beta0 (default: 1)}

\item{init_nu_TVS}{Initial value for nu (default: 1)}

\item{init_gamma_TVS}{Initial value for gamma (default: 1)}

\item{init_theta_TVS}{Initial value for theta (default: 0.5)}

\item{SS_t0_TVS}{Initial temperature for simulated annealing (default: 10.0)}

\item{SS_t1_TVS}{Final temperature for simulated annealing (default: 1.0)}

\item{hyper_mu_beta0_TVS}{Hyperparameter for beta0 mean (default: 0)}

\item{hyper_sigma_beta0_TVS}{Hyperparameter for beta0 variance (default: 1e6)}

\item{hyper_mu_nu_TVS}{Hyperparameter for nu mean (default: 1)}

\item{hyper_sigma_nu_TVS}{Hyperparameter for nu variance (default: 1)}

\item{hyper_c_gamma_TVS}{Hyperparameter for gamma shape (default: 0.0001)}

\item{hyper_d_gamma_TVS}{Hyperparameter for gamma rate (default: 0.0001)}

\item{hyper_a_theta_TVS}{Hyperparameter for theta shape (default: 1)}

\item{hyper_b_theta_TVS}{Hyperparameter for theta shape (default: -1.0)}

\item{max_iter_TVS}{Maximum number of iterations (default: 100)}

\item{tol_TVS}{Convergence tolerance (default: 1e-6)}

\item{add_correc_CiS}{Correction factor for CiS (default: 0.001)}
}
\value{
Vector of TVS values
}
\description{
Calculate TVS values
}
