% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{TVS_EM}
\alias{TVS_EM}
\title{EM algorithm for TVS}
\usage{
TVS_EM(
  dataXY,
  init_beta = rep(1, 8),
  init_beta0 = 1,
  init_nu = 1,
  init_gamma = 1,
  init_theta = 0.5,
  SS_t0 = 10,
  SS_t1 = 1,
  hyper_mu_beta0 = 0,
  hyper_sigma_beta0 = 1e+06,
  hyper_mu_nu = 1,
  hyper_sigma_nu = 1,
  hyper_c_gamma = 1e-04,
  hyper_d_gamma = 1e-04,
  hyper_a_theta = 1,
  hyper_b_theta = -1,
  max_iter = 100,
  tol = 1e-06,
  conv_type = "param"
)
}
\arguments{
\item{dataXY}{List containing data X and Y}

\item{init_beta}{Initial vector for beta}

\item{init_beta0}{Initial value for beta0 (default: 1)}

\item{init_nu}{Initial value for nu (default: 1)}

\item{init_gamma}{Initial value for gamma (default: 1)}

\item{init_theta}{Initial value for theta (default: 0.5)}

\item{SS_t0}{Initial temperature for simulated annealing (default: 10.0)}

\item{SS_t1}{Final temperature for simulated annealing (default: 1.0)}

\item{hyper_mu_beta0}{Hyperparameter for beta0 mean (default: 0)}

\item{hyper_sigma_beta0}{Hyperparameter for beta0 variance (default: 1e6)}

\item{hyper_mu_nu}{Hyperparameter for nu mean (default: 1)}

\item{hyper_sigma_nu}{Hyperparameter for nu variance (default: 1)}

\item{hyper_c_gamma}{Hyperparameter for gamma shape (default: 0.0001)}

\item{hyper_d_gamma}{Hyperparameter for gamma rate (default: 0.0001)}

\item{hyper_a_theta}{Hyperparameter for theta shape (default: 1)}

\item{hyper_b_theta}{Hyperparameter for theta shape (default: -1.0)}

\item{max_iter}{Maximum number of iterations (default: 100)}

\item{tol}{Convergence tolerance (default: 1e-6)}

\item{conv_type}{Convergence type (default: "param")}
}
\value{
List of results from the EM algorithm
}
\description{
EM algorithm for TVS
}
