% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{TVS_group}
\alias{TVS_group}
\title{TVS variable selection for a group of predictors}
\usage{
TVS_group(
  test_indices,
  dataXY,
  init_beta_TVS = rep(1, 8),
  B = 300,
  sig_cutoff = 0.05,
  init_beta0_TVS = 1,
  init_nu_TVS = 1,
  init_gamma_TVS = 1,
  init_theta_TVS = 0.5,
  SS_t0_TVS = 10,
  SS_t1_TVS = 1,
  hyper_mu_beta0_TVS = 0,
  hyper_sigma_beta0_TVS = 1e+06,
  hyper_mu_nu_TVS = 1,
  hyper_sigma_nu_TVS = 1,
  hyper_c_gamma_TVS = 1e-04,
  hyper_d_gamma_TVS = 1e-04,
  hyper_a_theta_TVS = 1,
  hyper_b_theta_TVS = -1,
  max_iter_TVS = 100,
  tol_TVS = 1e-06,
  add_correc_CiS = 0.001,
  beta_method = c("nlm", "optim", "maxLik", "cd_nlm", "cd_maxLik")
)
}
\arguments{
\item{test_indices}{Indices of predictors to test}

\item{dataXY}{List containing data X and Y}

\item{init_beta_TVS}{Initial vector for beta (default: rep(1, number of predictors))}

\item{B}{Number of iterations (default: 300)}

\item{sig_cutoff}{Threshold used for significance cutoff in variable selection (default: 0.05)}

\item{init_beta0_TVS}{Initial value for beta0 (default: 1)}

\item{init_nu_TVS}{Initial value for nu (default: 1)}

\item{init_gamma_TVS}{Initial value for gamma (default: 1)}

\item{init_theta_TVS}{Initial value for theta (default: 0.5)}

\item{SS_t0_TVS}{the "spike" hyperparameter in beta's spike-and-slab prior (default: 10)}

\item{SS_t1_TVS}{the "slab" hyperparameter in beta's spike-and-slab prior (default: 1)}

\item{hyper_mu_beta0_TVS}{the mean hyperparameter in beta0's Normal prior (default: 0)}

\item{hyper_sigma_beta0_TVS}{the variance hyperparameter in beta0's Normal prior (default: 1e6)}

\item{hyper_mu_nu_TVS}{the location hyperparameter in nu's log-normal prior (default: 1)}

\item{hyper_sigma_nu_TVS}{the scale hyperparameter in nu's log-normal prior (default: 1)}

\item{hyper_c_gamma_TVS}{the shape hyperparameter in gamma's gamma prior (default: 1e-4)}

\item{hyper_d_gamma_TVS}{the rate hyperparameter in gamma's gamma prior (default: 1e-4)}

\item{hyper_a_theta_TVS}{the shape hyperparameter a in theta's beta prior (default: 1)}

\item{hyper_b_theta_TVS}{the shape hyperparameter b in theta's beta prior (default: the number of predictors)}

\item{max_iter_TVS}{Maximum number of iterations (default: 100)}

\item{tol_TVS}{Convergence tolerance (default: 1e-6)}

\item{add_correc_CiS}{Correction factor for CiS (default: 1e-3)}

\item{beta_method}{Optimization method for beta updates ("nlm", "optim", "maxLik", "cd_nlm", "cd_maxLik"). Defaults to "nlm".}
}
\value{
List containing variable selection results
}
\description{
TVS variable selection for a group of predictors
}
