% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{TVS_multi_stage}
\alias{TVS_multi_stage}
\title{Multi-stage TVS algorithm}
\usage{
TVS_multi_stage(
  wrapper_beta,
  wrapper_beta0,
  wrapper_nu,
  wrapper_gamma,
  dataXY,
  init_beta_TVS = rep(1, 8),
  group_B = 20,
  indiv_B = 20,
  B_final = 300,
  group_cutoff = 3/20,
  indiv_cutoff = 3/20,
  group_size = 4,
  init_beta0_TVS = 1,
  init_nu_TVS = 1,
  init_gamma_TVS = 1,
  init_theta_TVS = 0.5,
  SS_t0_TVS = 10,
  SS_t1_TVS = 1,
  hyper_mu_beta0_TVS = 0,
  hyper_sigma_beta0_TVS = 1e+06,
  hyper_mu_nu_TVS = 1,
  hyper_sigma_nu_TVS = 1,
  hyper_c_gamma_TVS = 1e-04,
  hyper_d_gamma_TVS = 1e-04,
  hyper_a_theta_TVS = 1,
  hyper_b_theta_TVS = -1,
  max_iter_TVS = 100,
  tol_TVS = 1e-06,
  add_correc_CiS = 0.001
)
}
\arguments{
\item{wrapper_beta}{Function wrapper for beta}

\item{wrapper_beta0}{Function wrapper for beta0}

\item{wrapper_nu}{Function wrapper for nu}

\item{wrapper_gamma}{Function wrapper for gamma}

\item{dataXY}{List containing data X and Y}

\item{init_beta_TVS}{Initial vector for beta}

\item{group_B}{Number of iterations for group stage (default: 20)}

\item{indiv_B}{Number of iterations for individual stage (default: 20)}

\item{B_final}{Number of iterations for final stage (default: 300)}

\item{group_cutoff}{Cutoff value for group stage (default: 3.0/20.0)}

\item{indiv_cutoff}{Cutoff value for individual stage (default: 3.0/20.0)}

\item{group_size}{Size of groups (default: 4)}

\item{init_beta0_TVS}{Initial value for beta0 (default: 1)}

\item{init_nu_TVS}{Initial value for nu (default: 1)}

\item{init_gamma_TVS}{Initial value for gamma (default: 1)}

\item{init_theta_TVS}{Initial value for theta (default: 0.5)}

\item{SS_t0_TVS}{Initial temperature for simulated annealing (default: 10.0)}

\item{SS_t1_TVS}{Final temperature for simulated annealing (default: 1.0)}

\item{hyper_mu_beta0_TVS}{Hyperparameter for beta0 mean (default: 0)}

\item{hyper_sigma_beta0_TVS}{Hyperparameter for beta0 variance (default: 1e6)}

\item{hyper_mu_nu_TVS}{Hyperparameter for nu mean (default: 1)}

\item{hyper_sigma_nu_TVS}{Hyperparameter for nu variance (default: 1)}

\item{hyper_c_gamma_TVS}{Hyperparameter for gamma shape (default: 0.0001)}

\item{hyper_d_gamma_TVS}{Hyperparameter for gamma rate (default: 0.0001)}

\item{hyper_a_theta_TVS}{Hyperparameter for theta shape (default: 1)}

\item{hyper_b_theta_TVS}{Hyperparameter for theta shape (default: -1.0)}

\item{max_iter_TVS}{Maximum number of iterations (default: 100)}

\item{tol_TVS}{Convergence tolerance (default: 1e-6)}

\item{add_correc_CiS}{Correction factor for CiS (default: 0.001)}
}
\value{
List of multi-stage TVS results
}
\description{
Multi-stage TVS algorithm
}
