% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrapper_beta}
\alias{wrapper_beta}
\title{Wrapper function for beta optimization}
\usage{
wrapper_beta(
  beta,
  beta0,
  nu,
  gamma,
  betaPRE,
  t0 = 10,
  t1 = 1,
  Y_lk,
  X_lk,
  theta_lk = -2
)
}
\arguments{
\item{beta}{Initial beta parameter}

\item{beta0}{Beta0 parameter}

\item{nu}{Nu parameter}

\item{gamma}{Gamma parameter}

\item{betaPRE}{Previous beta value}

\item{t0}{Initial temperature for simulated annealing (default: 10)}

\item{t1}{Final temperature for simulated annealing (default: 1)}

\item{Y_lk}{Response vector}

\item{X_lk}{Predictor matrix}

\item{theta_lk}{Theta parameter (default: -2)}
}
\value{
Optimized beta parameter
}
\description{
Wrapper function for beta optimization
}
