% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrapper_beta0}
\alias{wrapper_beta0}
\title{Wrapper function for beta0 optimization}
\usage{
wrapper_beta0(
  beta0.lk,
  beta.lk,
  nu.lk,
  gamma.lk,
  hyper.mu.beta0 = 0,
  hyper.sigma.beta0 = 10^6,
  Y.lk,
  X.lk
)
}
\arguments{
\item{beta0.lk}{Initial beta0 parameter}

\item{beta.lk}{Beta parameter}

\item{nu.lk}{Nu parameter}

\item{gamma.lk}{Gamma parameter}

\item{hyper.mu.beta0}{Hyperparameter for beta0 mean (default: 0)}

\item{hyper.sigma.beta0}{Hyperparameter for beta0 variance (default: 10^6)}

\item{Y.lk}{Response vector}

\item{X.lk}{Predictor matrix}
}
\value{
Optimized beta0 parameter
}
\description{
Wrapper function for beta0 optimization
}
