% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrapper_beta}
\alias{wrapper_beta}
\title{Dispatcher for beta optimization}
\usage{
wrapper_beta(
  beta,
  beta0,
  nu,
  gamma,
  betaPRE,
  t0,
  t1,
  Y_lk,
  X_lk,
  theta_lk,
  method = getOption("TDVS_beta_method", "nlm")
)
}
\arguments{
\item{method}{Estimation function for beta (default: Quasi-Newton method implemented by R optim() function)}
}
\value{
Optimized beta parameter
}
\description{
Dispatcher for beta optimization
}
