% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrapper_beta}
\alias{wrapper_beta}
\title{Dispatcher for beta optimization}
\usage{
wrapper_beta(..., method = getOption("TVS_beta_method", "maxLik"))
}
\arguments{
\item{method}{Estimation function for beta (default: Quasi-Newton method implemented by R optim() function)}
}
\value{
Optimized beta parameter
}
\description{
Dispatcher for beta optimization
}
