% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrapper_gamma}
\alias{wrapper_gamma}
\title{Wrapper function for gamma optimization}
\usage{
wrapper_gamma(gamma, nu, error_lk, hyper_c, hyper_d)
}
\arguments{
\item{gamma}{Initial gamma parameter}

\item{nu}{Nu parameter}

\item{error_lk}{Error vector}

\item{hyper_c}{the shape hyperparameter in gamma's gamma prior (default: 10e-4)}

\item{hyper_d}{the rate hyperparameter in gamma's gamma prior (default: 10e-4)}
}
\value{
Optimized gamma parameter
}
\description{
Wrapper function for gamma optimization
}
